proper split of functionalities / wip downward compatibillity
